LayoutTests/fast/parser/block-nesting-cap-expected.txt

Test that the HTML parser does not allow the nesting depth of "block-level" elements to exceed 2048.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d2048.parentNode === d2046 is true
PASS d2047.parentNode === d2046 is true
PASS d2048.previousSibling === d2047 is true
PASS successfullyParsed is true

TEST COMPLETE
